fix(examples): correct eight defects in the Next.js session example
A review of the example as first pushed found eight problems, one of which
would have broken every signed-in visitor for anyone who copied it.

- `app/providers.tsx` seeded the provider with `initialToken` and no
  `initialSessionId`. The provider requires the pair and throws without it,
  so every signed-in visitor got a 500. Only `getInitialToken()` produces
  the session id, and it is unreachable from a page render — the seed is
  dropped, with the reason written down inline.
- `app/layout.tsx` read cookies purely to build that seed, which opted the
  whole tree — including `/callback` — out of static prerendering for
  nothing. It is a plain Server Component again.
- `middleware.ts` is deprecated in Next 16; it is now `proxy.ts`.
- The proxy ran on every request, including asset and API fetches, spending
  a token mint on requests that can never use one. It is gated on
  `Sec-Fetch-Dest: document`.
- The proxy copied only the first `Set-Cookie` and dropped every other
  header, losing `Cache-Control: no-store` on a response carrying a
  credential. It forwards `getSetCookie()` entry by entry and preserves the
  rest.
- `server/logto-cookie.ts` read the app origin from `NEXT_PUBLIC_APP_ORIGIN`,
  a build-time-inlined variable, for a server-only value used to validate
  redirects. It reads `APP_ORIGIN`.
- `app/dashboard.tsx` rendered "no identity on the server" for a signed-out
  visitor, which describes the mechanism rather than the state, and its
  `usePreloadedQuery` comment described the wrong thing.
- The README table, the org-scopes snippet (no imports, no note that scopes
  only take effect after a fresh sign-in) and the production `APP_ORIGIN`
  guidance were all wrong or incomplete.

`docs/content/docs/nextjs.mdx` follows: middleware → proxy, the
`Sec-Fetch-Dest` gate, the header-forwarding rule, and a paragraph on why
the provider's SSR seed does not apply in the App Router.

The lockfile gains only the new importer: main's peer-resolution keys are
left alone so the diff carries nothing unrelated.
