+++ examples/nextjs-session/.env.example
@@ -3,8 +3,12 @@
 NEXT_PUBLIC_CONVEX_URL=
 
 # The origin this app is served from. The cookie handler accepts requests only
-# from exact origins in this list — never a wildcard.
-NEXT_PUBLIC_APP_ORIGIN=http://localhost:3000
+# from this exact origin — never a wildcard.
+#
+# Deliberately not NEXT_PUBLIC_: Next inlines those at build time, so a public
+# name set in the production environment would be ignored in favour of whatever
+# the build machine had, and every request to the handler would be answered 403.
+APP_ORIGIN=http://localhost:3000
 
 # Nothing about Logto belongs here. Session mode keeps the endpoint, the app id
 # and the client secret on the Convex deployment, so the bundle carries none of
