+++ examples/nextjs-session/README.md
@@ -15,22 +15,37 @@ read.
 |---|---|
 | `app/api/logto/[route]/route.ts` | The five-line catch-all mount. App Router handlers are standard `Request`/`Response`, so the handler needs no adapter. |
 | `server/logto-cookie.ts` | One handler configuration shared by the route and the middleware, with `idTokenCookie: true`. |
-| `app/layout.tsx` | A Server Component reading the ID token cookie and seeding the provider, so the first client render agrees with the server's. |
-| `app/page.tsx` | `preloadQuery` with that token — authenticated HTML, no loading flash. |
-| `middleware.ts` | Where `getInitialToken()` belongs, because middleware can set cookies. |
+| `app/page.tsx` | `readLogtoIdTokenCookie` + `preloadQuery` — the server renders the real identity instead of a spinner. |
+| `proxy.ts` | Where `getInitialToken()` belongs, because a proxy can set cookies — gated to document requests. |
 | `app/dashboard.tsx` | Device list, live revocation, organization permissions, and the live Logto profile. |
 
 ## The one rule worth reading before you copy this
 
 **Never call `getInitialToken()` from a layout or a page.** It rotates the
 session cookie, and a Server Component cannot set cookies — Next.js allows that
-only in Route Handlers, Server Actions, and middleware. The rotated cookie would
+only in Route Handlers, Server Actions, and the proxy. The rotated cookie would
 be dropped, the browser would keep presenting a superseded token, and once it
 fell outside the reuse window the component would read it as theft and kill the
 session.
 
-Read `readLogtoIdTokenCookie` in a page; call `getInitialToken()` in middleware.
-This example does exactly that, and `middleware.ts` says why inline.
+Read `readLogtoIdTokenCookie` in a page; call `getInitialToken()` in the proxy.
+This example does exactly that, and `proxy.ts` says why inline.
+
+Two consequences worth understanding before you copy the wiring:
+
+- **Rotate once per document request.** Every `getInitialToken()` call is a real
+  Logto round trip that rotates the session token. A matcher that also catches
+  `/favicon.ico`, images and RSC prefetches fires several for one page view, and
+  whichever `Set-Cookie` the browser keeps last may be an older generation than
+  the server's — which the next client refresh presents outside its reuse window
+  and the component correctly reads as theft. `proxy.ts` gates on
+  `Sec-Fetch-Dest: document`.
+- **The provider gets no SSR seed here.** It takes `initialToken` and
+  `initialSessionId` as a pair, and the session id comes only from
+  `getInitialToken()` — which cannot run in a render. So the server renders the
+  identity itself (`app/page.tsx`) and the client establishes its own auth on
+  mount; the first paint is authenticated, and the client catches up a moment
+  later without the page changing shape.
 
 ## Setup
 
@@ -49,20 +64,30 @@ npx convex env set LOGTO_CLIENT_SECRET your-traditional-web-app-secret
 pnpm dev
 ```
 
-To see organization permissions do anything, add the organization scopes on the
-Convex side — they are server-configured, because the browser cannot request its
-own:
+To see the organization buttons do anything, ask for the organization scopes on
+the Convex side — they are server-configured, because the browser cannot request
+its own:
 
 ```ts title="convex/auth.ts"
-logtoSessionApi(components.logto, {
-  scopes: [ORGANIZATIONS_SCOPE, ORGANIZATION_ROLES_SCOPE],
-});
+import { ORGANIZATIONS_SCOPE, ORGANIZATION_ROLES_SCOPE, logtoSessionApi } from "convex-logto";
+
+export const { /* ...the eleven exports... */ } = logtoSessionApi(
+  components.logto,
+  { scopes: [ORGANIZATIONS_SCOPE, ORGANIZATION_ROLES_SCOPE] },
+);
 ```
+
+**Sign out and back in afterwards.** Scopes are fixed at authorization time and
+a grant cannot be widened in place, so an existing session's ID token keeps its
+old claims and the buttons stay disabled until a new sign-in.
 
 ## Going to production
 
-- `NEXT_PUBLIC_APP_ORIGIN` must be your real origin. `allowedOrigins` takes
-  exact origins and rejects wildcards.
+- `APP_ORIGIN` must be your real origin, and it is deliberately *not* a
+  `NEXT_PUBLIC_` name: Next inlines those at build time, so a public one set in
+  the production environment would be ignored in favour of the build machine's
+  and every request to the handler would be answered `403`. `allowedOrigins`
+  takes exact origins and rejects wildcards.
 - The cookies are `__Host-` prefixed and `Secure`, so everything except
   `localhost` needs HTTPS.
 - A cookie rides on **every** same-origin request, so the ID token reaches
